backend/app/schemas/auth.py
"""Auth schemas."""

from datetime import datetime

from pydantic import BaseModel, EmailStr, Field


class RegisterRequest(BaseModel):
    username: str = Field(min_length=3, max_length=50)
    email: EmailStr
    password: str = Field(min_length=6, max_length=128)
    full_name: str = Field(min_length=1, max_length=100)


class LoginRequest(BaseModel):
    username: str
    password: str


class TokenResponse(BaseModel):
    access_token: str
    token_type: str = "bearer"


class UserResponse(BaseModel):
    id: int
    username: str
    email: str
    full_name: str
    avatar_url: str | None = None
    bio: str | None = None
    is_active: bool
    is_admin: bool
    is_superadmin: bool = False
    created_at: datetime

    model_config = {"from_attributes": True}
